# src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: Java batch API
  description: developed java batch
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /batch/start:
    post:
      tags:
      - batch-controller
      operationId: startBatch
      responses:
        "200":
          description: OK
  /batch/createUsers:
    post:
      tags:
      - batch-controller
      operationId: saveUsers
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /batch/createUser:
    post:
      tags:
      - batch-controller
      operationId: saveUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
  /batch/createProfile:
    post:
      tags:
      - batch-controller
      operationId: saveProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Profile'
  /batch/profiles:
    get:
      tags:
      - batch-controller
      operationId: getProfiles
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'
  /batch/Users:
    get:
      tags:
      - batch-controller
      operationId: getUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        emailVerified:
          type: boolean
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
        createdDate:
          type: string
          format: date-time
        lastModifiedDate:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        id:
          type: integer
          format: int32
        fullName:
          type: string
        userId:
          type: integer
          format: int32
        createdDate:
          type: string
          format: date-time
        lastModifieDate:
          type: string
          format: date-time